shared_scripts/stream_sketcher/requirements.txt
aiohttp>=3.9
PyYAML>=6.0
pyarrow
orjson
duckdb
